Added Inspect Functionality
NEO now inspects and memorizes objects. Tries not to inspect the same
object twice. Data is saved between runs, so he will not approach the
apple after the first run unless the .db file is deleted.

Added Object_Meta class to overcome the fact that the Object class
cannot be deepcopy'd.

Made detected_objects a lsit so it is iterable.

Added position columns to the SQL DB.