Fixed DB Error, Rotated Raycasts with NEO
The DB error wasn't a typo - the old DB just had to be deleted.

Added 2 more objects to the map to show a problem we can fix.

Fixed legs.py so he doesn't turn when you press the arrow keys.

Made the raycasts rotate at the same angle as NEO. Can be made
horizontal.